Refactor cluster local logic
This is in support of #31401
