+++ pilot/pkg/xds/ep_filters_test.go
@@ -59,6 +59,7 @@ func TestEndpointsByNetworkFilter(t *testing.T) {
 	//  - 1 gateway for network3
 	//  - 0 gateways for network4
 	env := environment()
+	env.Init()
 
 	// Test endpoints creates:
 	//  - 2 endpoints in network1
@@ -229,6 +230,7 @@ func TestEndpointsByNetworkFilter_SkipLBWithHostname(t *testing.T) {
 	serviceDiscovery.SetGatewaysForNetwork("network2", &model.Gateway{Addr: "aeiou.scooby.do", Port: 80})
 
 	env.ServiceDiscovery = serviceDiscovery
+	env.Init()
 
 	// Test endpoints creates:
 	//  - 2 endpoints in network1
